Add tha admin authorization to the delete route in backend

+++ backend/controllers/pagePosts.js
@@ -8,7 +8,17 @@ exports.createPagePost = (req, res, next) => {
     // delete req.body._id;
     //console.log()
     const pagePost = new PagePost({
-      ...req.body
+      text: req.body.text,
+      title: req.body.title,
+    
+      author: { 
+          id : req.auth.userId,
+          name : req.auth.userName,
+          admin : req.auth.userAdmin
+          // firstName : {type : mongoose.Schema.Types.String },
+          // lastName : {type : mongoose.Schema.Types.String },
+
+      },
     });
     pagePost.save()
       .then(() => res.status(201).json(pagePost)) // try to console.log(res) if error comes back
@@ -22,10 +32,32 @@ exports.createPagePost = (req, res, next) => {
 
   };
 
-exports.deletePagePost = (req, res, next) => {
-    pagePosts.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-        .catch(error => res.status(400).json({ error }));
+exports.deletePagePost = async (req, res, next) => {
+  try {
+    
+    const post = await PagePost.findById(req.params.id)
+    const postAuthor = post.author.id
+    const postAdmin = post.author.admin
+    const userId = req.auth.userId
+    const admin = req.auth.userAdmin
+
+    if (postAuthor === userId || admin) {
+        pagePosts.deleteOne({ _id: req.params.id })
+          .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
+          .catch(error => res.status(400).json({ error }));
+    } else {
+      pagePosts.findOne({ _id: req.params.id })
+          .then(() => res.status(200).json({ message: 'You are not authorised to delete this post !'}))
+          .catch(error => res.status(400).json({ error }));
+    }
+
+    console.log ("this is my information", post, userId, admin, postAuthor, postAdmin)
+    
+  } catch (err) {
+    console.log (err)
+  }
+
+    
 }
 
 exports.updatePagePost = (req, res, next) => {
